pyproject.toml
[tool.setuptools.packages.find]
include = ["agent*", "mcp_servers*"]

[project]
name = "localops-agent"
version = "0.1.0"
description = "Local-first work OS: MCP-connected agent with approval-based execution"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.9",
    "fastmcp>=2.0.0",
    "openai>=1.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",       # 터미널 UI (승인 프롬프트 등)
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
